Player: Get 1x1 gaps to work by nearest collision (time)
The first intersection time is now calculated and used
instead of judging by the distance of the intersection.
As a result, collisions pointing into the opposite
movement direction are ignored to avoid false-positives.

+++ src/core/utils.h
@@ -31,7 +31,7 @@ void sleep_ms(long delay);
 
 // ------------------ Numeric ------------------
 
-inline float get_sign(float f)
+inline int get_sign(float f)
 {
 	if (f > 0.0001f)
 		return 1;
